Make `num_cards` int instead of size_t

+++ day-11a-cf-424-b.cpp
@@ -127,7 +127,7 @@ public:
 
 int main()
 {
-    size_t num_cards;
+    int num_cards;
 
     std::cin >> num_cards;
